APP-3419 The PlanYear tests were failing with deprecation warnings after I added the InsertAttributeStripper to them (#35)
There's a frustrating (to the whole rails community) issue with rails, the ruby TImeout library, and how it and return, throw, and catch interact which means for the moment return statements inside of transaction blocks are recommended against and will probably start throwing errors in production environments in the next rails version. They're currently throwing deprecation warnings which are causing tests to fail. This removes the return statements and restructures the InsertAttributeStripper to return the correct value without them.